feat: 대기열 구현 완성

+++ queue/src/main/java/com/example/queue/QueueDTO.java
@@ -2,6 +2,6 @@ package com.example.queue;
 
 public record QueueDTO(
         String userId,
-        int percent
+        double percent
 ) {
 }
